+++ casatables/src/glue.h
@@ -16,7 +16,8 @@
 #ifndef CASA_TYPES_ALREADY_DECLARED
 
 // copied from casa/Utilities/DataType.h:
-typedef enum GlueDataType {
+typedef enum GlueDataType
+{
     TpBool,
     TpChar,
     TpUChar,
@@ -52,6 +53,7 @@ typedef enum GlueDataType {
 
 typedef struct GlueTable GlueTable;
 typedef struct GlueTableRow GlueTableRow;
+typedef struct GlueTableDesc GlueTableDesc;
 #endif
 
 // OMG, strings. First of all: casacore::String is a subclass of std::string,
@@ -66,12 +68,14 @@ typedef struct GlueTableRow GlueTableRow;
 // to use C++->Rust callbacks to be able to copy string contents before they
 // are deallocated at the C++ layer.
 
-typedef struct StringBridge {
+typedef struct StringBridge
+{
     const void *data;
     unsigned long n_bytes;
 } StringBridge;
 
-typedef struct ExcInfo {
+typedef struct ExcInfo
+{
     char message[512];
 } ExcInfo;
 
@@ -83,15 +87,67 @@ typedef void (*StringBridgeCallback)(const StringBridge *name, void *ctxt);
 // additional information we'd like to to transfer.
 typedef void (*KeywordInfoCallback)(const StringBridge *name, GlueDataType dtype, void *ctxt);
 
-typedef enum TableOpenMode {
+typedef enum TableOpenMode
+{
     TOM_OPEN_READONLY = 1,
     TOM_OPEN_RW = 2,
     TOM_CREATE = 3,
 } TableOpenMode;
 
-extern "C" {
+typedef enum TableCreateMode
+{
+    // create table
+    TCM_NEW = 1,
+    // create table (may not exist)
+    TCM_NEW_NO_REPLACE = 2,
+    // New table, which gets marked for delete"
+    TCM_SCRATCH = 3,
+} TableCreateMode;
+
+typedef enum TableDescCreateMode
+{
+    //    Create a new table description file.
+    //    The TableDesc destructor will write the table description into the file.
+    TDM_NEW,
+    //    As option New, but an exception will be thrown if the table
+    //    description file already exists.
+    TDM_NEW_NO_REPLACE,
+    //    Create a temporary table description. The table description will
+    //    be lost when the TableDesc object is destructed.
+    //    This is useful to create a Table object without storing the
+    //    description separately.
+    //    Note that the Table object maintains its own description (i.e. it
+    //    copies the description when being constructed).
+    TDM_SCRATCH,
+    // Note, there are other options here which could be enabled if needed.
+    //  <li> Old
+    //    Open an existing table description file as readonly.
+    //  <li> Update
+    //    Open an existing table description file as read/write
+    //    The TableDesc destructor will rewrite the possibly changed
+    //    description.
+    //  <li> Delete
+    //    Delete the table description file. This gets done by the destructor.
+} TableDescOption;
+
+extern "C"
+{
     int data_type_get_element_size(const GlueDataType ty);
 
+    GlueTableDesc *tabledesc_create(const StringBridge &type, const TableDescCreateMode mode, 
+                                    ExcInfo &exc);
+    GlueTableDesc *tabledesc_add_scalar_column(GlueTableDesc &table_desc, GlueDataType data_type, 
+                                               const StringBridge &col_name, const StringBridge 
+                                               &comment, bool direct, bool undefined, ExcInfo &exc);
+    GlueTableDesc *tabledesc_add_array_column(GlueTableDesc &table_desc, GlueDataType data_type, 
+                                              const StringBridge &col_name, const StringBridge &comment, 
+                                              bool direct, bool undefined, ExcInfo &exc);
+    GlueTableDesc *tabledesc_add_fixed_array_column(GlueTableDesc &table_desc, GlueDataType data_type, 
+                                                    const StringBridge &col_name, const StringBridge &comment, 
+                                                    const unsigned long n_dims, const unsigned long *dims, 
+                                                    bool direct, bool undefined, ExcInfo &exc);
+    GlueTable *table_create(const StringBridge &path, GlueTableDesc &table_desc, 
+                            unsigned long n_rows, const TableCreateMode mode, ExcInfo &exc);
     GlueTable *table_alloc_and_open(const StringBridge &path, const TableOpenMode mode, ExcInfo &exc);
     void table_close_and_free(GlueTable *table, ExcInfo &exc);
     unsigned long table_n_rows(const GlueTable &table);
@@ -101,6 +157,8 @@ extern "C" {
     unsigned long table_n_keywords(const GlueTable &table);
     int table_get_keyword_info(const GlueTable &table, KeywordInfoCallback callback,
                                void *ctxt, ExcInfo &exc);
+    int table_put_keyword(GlueTable &table, const StringBridge &kw_name, const GlueDataType data_type, 
+                          void *data, ExcInfo &exc);
     int table_copy_rows(const GlueTable &source, GlueTable &dest, ExcInfo &exc);
     int table_deep_copy_no_rows(const GlueTable &table, const StringBridge &dest_path, ExcInfo &exc);
     int table_get_column_info(const GlueTable &table, const StringBridge &col_name,
@@ -108,6 +166,13 @@ extern "C" {
                               int *is_scalar, int *is_fixed_shape, int *n_dim,
                               unsigned long dims[8], ExcInfo &exc);
     int table_remove_column(GlueTable &table, const StringBridge &col_name, ExcInfo &exc);
+    int table_add_scalar_column(GlueTable &table, GlueDataType data_type, const StringBridge &col_name, 
+                                const StringBridge &comment, bool direct, bool undefined, ExcInfo &exc);
+    int table_add_array_column(GlueTable &table, GlueDataType data_type, const StringBridge &col_name, 
+                               const StringBridge &comment, bool direct, bool undefined, ExcInfo &exc);
+    int table_add_fixed_array_column(GlueTable &table, GlueDataType data_type, const StringBridge &col_name, 
+                                     const StringBridge &comment, const unsigned long n_dims, 
+                                     const unsigned long *dims, bool direct, bool undefined, ExcInfo &exc);
     int table_get_scalar_column_data(const GlueTable &table, const StringBridge &col_name,
                                      void *data, ExcInfo &exc);
     int table_get_scalar_column_data_string(const GlueTable &table, const StringBridge &col_name,