+++ miriad/CHANGELOG.md
@@ -1,3 +1,7 @@
+# rubbl_miriad 0.2.5 (2021-10-07)
+
+- build(deps): bump openat from 0.1.20 to 0.1.21
+
 # rubbl_miriad 0.2.4 (2021-04-01)
 
 - Remove a bunch of superfluous dependencies